refactor: extract components from problem set and problem page
