+++ src/components/index.ts
@@ -1,2 +1,5 @@
 export { default as MyCollections } from './MyCollections.vue';
 export { default as MyProfile } from './MyProfile.vue';
+export { default as ProblemTable } from './ProblemTable.vue';
+export { default as RecommendForm } from './RecommendForm.vue';
+export { default as RecommendList } from './RecommendList.vue';
